* x64 solution fixes. * x32-based logonservers can now link to x64-based ascents, and vice-versa.
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2194 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/MiscHandler.cpp
@@ -270,7 +270,7 @@ void WorldSession::HandleLootMoneyOpcode( WorldPacket & recv_data )
 			if(!targets.size())
 				return;
 
-			uint32 share = money / targets.size();
+			uint32 share = money / uint32(targets.size());
 
 			pkt.SetOpcode(SMSG_LOOT_MONEY_NOTIFY);
 			pkt << share;
